RAG-Chatbot/requirements.txt
llama-index
llama-index-vector-stores-chroma
llama-index-readers-file
llama-index-embeddings-huggingface
chromadb
unstructured[doc,docx,ppt,pptx,pdf,image]
langchain
chainlit
python-dotenv
fastapi
uvicorn[standard]
jinja2
unsloth
transformers==4.56.2
trl==0.22.2